+++ modules/__tests__/body-test.js
@@ -1,8 +1,8 @@
 import expect from 'expect'
 import { body } from '../index'
 
-const echo = (url, options) =>
-  Promise.resolve({ url, options })
+const echo = (input, options) =>
+  Promise.resolve({ input, options })
 
 describe('body', () => {
   it('sets the body of the request', () => {