+++ modules/__tests__/header-test.js
@@ -1,8 +1,8 @@
 import expect from 'expect'
 import { header } from '../index'
 
-const echo = (url, options) =>
-  Promise.resolve({ url, options })
+const echo = (input, options) =>
+  Promise.resolve({ input, options })
 
 describe('header', () => {
   it('sets the value of a request header', () =>