+++ modules/__tests__/method-test.js
@@ -1,8 +1,8 @@
 import expect from 'expect'
 import { method } from '../index'
 
-const echo = (url, options) =>
-  Promise.resolve({ url, options })
+const echo = (input, options) =>
+  Promise.resolve({ input, options })
 
 describe('method', () => {
   it('sets the request method', () =>